+++ api/src/actions/conversation/GET/ConversationByIdAction.php
@@ -1,30 +1,31 @@
 <?php
 
-namespace api\actions;
+namespace api\actions\conversation\GET;
 
 use Psr\Http\Message\ResponseInterface as Response;
 use Psr\Http\Message\ServerRequestInterface as Request;
 
 // SERVICE
-use api\services\UserService as UserService;
+use api\services\ConversationService as ConversationService;
 use api\services\utils\FormatterAPI;
 
 //Exception
 use api\errors\exceptions\RessourceNotFoundException as RessourceNotFoundException;
 use Slim\Exception\HttpNotFoundException;
 
-final class UserByIdAction
+final class ConversationByIdAction
 {
-  public function __invoke(Request $rq, Response $rs, array $args): Response {
-    $userService = new UserService();
+  public function __invoke(Request $rq, Response $rs, array $args): Response
+  {
+    $conversationService = new ConversationService();
     try {
-      $userById = $userService->getUserByID($args['id']);
+      $conversationById = $conversationService->getConversationByID($args['id']);
     } catch (RessourceNotFoundException  $e) {
       throw new HttpNotFoundException($rq, $e->getMessage());
     }
     $data = [
-      'Filtre' => 'UserById',
-      'Result' => $userById
+      'Filtre' => 'conversationById',
+      'Result' => $conversationById
     ];
 
     return FormatterAPI::formatResponse($rq, $rs, $data);
