+++ api/src/actions/user/GET/UserByIdAction.php
@@ -1,30 +1,31 @@
 <?php
 
-namespace api\actions\conversation;
+namespace api\actions\user\GET;
 
 use Psr\Http\Message\ResponseInterface as Response;
 use Psr\Http\Message\ServerRequestInterface as Request;
 
 // SERVICE
-use api\services\ConversationService as ConversationService;
+use api\services\UserService as UserService;
 use api\services\utils\FormatterAPI;
 
 //Exception
 use api\errors\exceptions\RessourceNotFoundException as RessourceNotFoundException;
 use Slim\Exception\HttpNotFoundException;
 
-final class ConversationByIdAction
+final class UserByIdAction
 {
-  public function __invoke(Request $rq, Response $rs, array $args): Response {
-    $conversationService = new ConversationService();
+  public function __invoke(Request $rq, Response $rs, array $args): Response
+  {
+    $userService = new UserService();
     try {
-      $conversationById = $conversationService->getConversationByID($args['id']);
+      $userById = $userService->getUserByID($args['id']);
     } catch (RessourceNotFoundException  $e) {
       throw new HttpNotFoundException($rq, $e->getMessage());
     }
     $data = [
-      'Filtre' => 'conversationById',
-      'Result' => $conversationById
+      'Filtre' => 'UserById',
+      'Result' => $userById
     ];
 
     return FormatterAPI::formatResponse($rq, $rs, $data);
